Not so OOP approach, v1.0 approach
I just finished reading a book from national author because of school and it was oriented in WW2 and stuff and it's called "Romeo, Juliet and darkness" and basically it's about 18yo student who lives in Czech Republic(my country) and falls in love with Jude girl. You can see where this is going... It ended tragically and it made me cry.. Really.. My first book with real emotions to it.. I mean, fuck me, it's just a book but I feel so fucking sorry for them and ahhh.. Just writing this makes me cry omg I need to stop.. For my emotional instability GitHub is gonna need to store so many useless bytes noone's ever gonna find || read.. But if it has translation to english (I think it has, it's really famous here) you should definitely read it! It's short but man, so damn hard to swallow...